Add cross-compilation support (WIP)
Also use conda-build 3's multiple sources feature.

WIP because I either need to remove the (commented out)
call to autoreconf or I need to comment it back in and
add autoconf as a build dependency.

The deciding factor here is whether the autoconf changes
around cross-compilation detection.  We need this to not
be detected as being cross-compiled as when that happens
tk assumes strtod is broken (as it assumes it cannot run
test executables) and builds its own implementation then
does not link to it (properly?).